move inheritance panel test
