+++ packages/host/tests/acceptance/operator-mode-test.ts
@@ -903,108 +903,6 @@ module('Acceptance | operator mode tests', function (hooks) {
         .dom('[data-test-code-mode-card-preview-body ] .edit-card')
         .exists();
     });
-
-    test('card inheritance panel will show json instance definition and module definition when is set to code', async function (assert) {
-      let operatorModeStateParam = stringify({
-        stacks: [
-          [
-            {
-              id: 'http://test-realm/test/Person/fadhlan',
-              format: 'isolated',
-            },
-          ],
-        ],
-        submode: 'code',
-        codePath: `http://test-realm/test/Person/fadhlan.json`,
-      })!;
-
-      await visit(
-        `/?operatorModeEnabled=true&operatorModeState=${encodeURIComponent(
-          operatorModeStateParam,
-        )}`,
-      );
-
-      await waitUntil(() => find('[data-test-card-inheritance-panel]'));
-      await waitUntil(() => find('[data-test-card-module-definition]'));
-      await waitUntil(() => find('[data-test-card-instance-definition]'));
-
-      assert.dom('[data-test-card-module-definition]').includesText('Person');
-      assert
-        .dom(
-          '[data-test-card-module-definition] [data-test-definition-file-extension]',
-        )
-        .includesText('.GTS');
-      assert
-        .dom(
-          '[data-test-card-module-definition] [data-test-definition-realm-name]',
-        )
-        .includesText('Test Workspace B');
-      assert
-        .dom('[data-test-card-module-definition]')
-        .doesNotHaveClass('active');
-      assert
-        .dom('[data-test-card-instance-definition]')
-        .includesText('Fadhlan');
-      assert
-        .dom(
-          '[data-test-card-instance-definition] [data-test-definition-file-extension]',
-        )
-        .includesText('.JSON');
-      assert
-        .dom(
-          '[data-test-card-instance-definition] [data-test-definition-realm-name]',
-        )
-        .includesText('Test Workspace B');
-      assert
-        .dom(
-          '[data-test-card-instance-definition] [data-test-definition-info-text]',
-        )
-        .includesText('Last saved was a few seconds ago');
-
-      assert.dom('[data-test-card-instance-definition]').hasClass('active');
-    });
-
-    test('clicking on listed card definition in the inheritance panel will display new definition and update card url', async function (assert) {
-      let operatorModeStateParam = stringify({
-        stacks: [
-          [
-            {
-              id: 'http://test-realm/test/Person/fadhlan',
-              format: 'isolated',
-            },
-          ],
-        ],
-        submode: 'code',
-        codePath: `http://test-realm/test/Person/fadhlan.json`,
-      })!;
-
-      await visit(
-        `/?operatorModeEnabled=true&operatorModeState=${encodeURIComponent(
-          operatorModeStateParam,
-        )}`,
-      );
-
-      await waitUntil(() => find('[data-test-card-inheritance-panel]'));
-      await waitUntil(() => find('[data-test-card-module-definition]'));
-      await waitUntil(() => find('[data-test-card-instance-definition]'));
-
-      assert.dom('[data-test-card-module-definition]').includesText('Person');
-      assert
-        .dom(
-          '[data-test-card-module-definition] [data-test-definition-file-extension]',
-        )
-        .includesText('.GTS');
-      assert
-        .dom('[data-test-card-url-bar-input]')
-        .hasValue(`${testRealmURL}Person/fadhlan.json`);
-      await click('[data-test-card-module-definition]');
-      await waitUntil(() => find('[data-test-card-module-definition]'));
-      assert
-        .dom('[data-test-card-url-bar-input]')
-        .hasValue(`${testRealmURL}person.gts`);
-      assert.dom('[data-test-card-module-definition]').includesText('Person');
-      assert.dom('[data-test-card-instance-definition]').doesNotExist();
-    });
   });
 
   test<TestContextWithSSE>('stack item live updates when index changes', async function (assert) {
